Move message send handling to a function
- Decreases copy-pasted code by A LOT
