+++ commands/clap.js
@@ -1,7 +1,5 @@
 const { SlashCommand, CommandOptionType } = require('slash-create');
-const fs = require('fs');
-const path = require("path");
-const SnowflakeCodon = require("snowflake-codon");
+const { sendMessage } = require('../utils/sendmessage.js')
 
 module.exports = class extends SlashCommand {
     constructor(creator) {
@@ -23,27 +21,11 @@ module.exports = class extends SlashCommand {
         try {
             const text = ctx.options.text;
             const convertedtext = text.replace(/ /g, "👏");
-            const generator = new SnowflakeCodon(1, 99, 2021, 200);
 
             await ctx.defer();
 
 
-            if (convertedtext.length > 2000) { // if converted text is too long to send in discord
-                var snowflakeid = generator.nextId();
-
-                fs.writeFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'), convertedtext); // write to file
-
-                ctx.sendFollowUp({
-                    content: "", file: {
-                        name: 'uwuify-' + snowflakeid + '.txt',
-                        file: fs.readFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'))
-                    }
-                });
-                fs.unlinkSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt')); // delete file
-                return;
-            }
-
-            ctx.sendFollowUp({ content: convertedtext });
+            sendMessage(convertedtext, ctx);
         } catch (error) {
             console.error(error);
         }
