+++ commands/emojify.js
@@ -1,8 +1,6 @@
 const { SlashCommand, CommandOptionType } = require('slash-create');
 const replaceWord = require('replace-word');
-const fs = require('fs');
-const path = require("path");
-const SnowflakeCodon = require("snowflake-codon");
+const { sendMessage } = require('../utils/sendmessage.js')
 
 module.exports = class extends SlashCommand {
     constructor(creator) {
@@ -30,7 +28,6 @@ module.exports = class extends SlashCommand {
         try {
             const text = ctx.options.text;
             var emojifiedtext = replaceWord.emojipasta(text, 100);
-            const generator = new SnowflakeCodon(1, 99, 2021, 200);          
 
             await ctx.defer();
 
@@ -40,22 +37,7 @@ module.exports = class extends SlashCommand {
                 emojifiedtext = replaceWord.emojipasta(text, ctx.options.density);
             };
 
-            if (emojifiedtext.length > 2000) { // if converted text is too long to send in discord
-                var snowflakeid = generator.nextId();
-
-                fs.writeFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'), emojifiedtext); // write to file
-
-                ctx.sendFollowUp({
-                    content: "", file: {
-                        name: 'uwuify-' + snowflakeid + '.txt',
-                        file: fs.readFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'))
-                    }
-                });
-                fs.unlinkSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt')); // delete file
-                return;
-            }
-
-            ctx.sendFollowUp({ content: emojifiedtext });
+            sendMessage(emojifiedtext, ctx);
         } catch (error) {
             console.error(error);
         }
