+++ commands/morse.js
@@ -1,8 +1,6 @@
 const { SlashCommand, CommandOptionType } = require('slash-create');
 const replaceWord = require('replace-word');
-const fs = require('fs');
-const path = require("path");
-const SnowflakeCodon = require("snowflake-codon");
+const { sendMessage } = require('../utils/sendmessage.js')
 
 module.exports = class extends SlashCommand {
     constructor(creator) {
@@ -24,27 +22,11 @@ module.exports = class extends SlashCommand {
         try {
             const text = ctx.options.text;
             const convertedtext = replaceWord.toMorse(text);
-            const generator = new SnowflakeCodon(1, 99, 2021, 200);
 
             await ctx.defer();
 
 
-            if (convertedtext.length > 2000) { // if converted text is too long to send in discord
-                var snowflakeid = generator.nextId();
-
-                fs.writeFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'), convertedtext); // write to file
-
-                ctx.sendFollowUp({
-                    content: "", file: {
-                        name: 'uwuify-' + snowflakeid + '.txt',
-                        file: fs.readFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'))
-                    }
-                });
-                fs.unlinkSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt')); // delete file
-                return;
-            }
-
-            ctx.sendFollowUp({ content: convertedtext });
+            sendMessage(convertedtext, ctx);
         } catch (error) {
             console.error(error);
         }
