+++ commands/uwuify.js
@@ -1,8 +1,6 @@
 const { SlashCommand, CommandOptionType } = require('slash-create');
 const { Uwuifier } = require('@patarapolw/uwuifier');
-const fs = require('fs');
-const path = require("path");
-const SnowflakeCodon = require("snowflake-codon");
+const { sendMessage } = require('../utils/sendmessage.js')
 
 module.exports = class extends SlashCommand {
     constructor(creator) {
@@ -35,7 +33,6 @@ module.exports = class extends SlashCommand {
     async run(ctx) {
         try {
             const uwuifier = new Uwuifier(); // create new uwuifier instance
-            const generator = new SnowflakeCodon(1, 99, 2021, 200);
             const text = ctx.options.text;
             var uwuifiedtext = uwuifier.uwuifyWords(text);
 
@@ -50,25 +47,7 @@ module.exports = class extends SlashCommand {
                 uwuifiedtext = uwuifier.uwuifyExclamations(uwuifiedtext); // it's quite shrimple
             };
 
-
-            if (uwuifiedtext.length > 2000) { // if converted text is too long to send in discord
-                var snowflakeid = generator.nextId();
-
-                fs.writeFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'), uwuifiedtext); // write to file
-
-                ctx.sendFollowUp({
-                    content: "", file: {
-                        name: 'uwuify-' + snowflakeid + '.txt',
-                        file: fs.readFileSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt'))
-                    }
-                });
-                fs.unlinkSync(path.resolve('./temp/uwuify-' + snowflakeid + '.txt')); // delete file
-                return;
-            } // :3
-
-
-            ctx.sendFollowUp({ content: uwuifiedtext });
-
+            sendMessage(uwuifiedtext, ctx);
         } catch (error) {
             console.error(error);
         }
