+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "uwuifier",
-    "version": "1.4.6",
+    "version": "1.5.0",
     "description": "The discord bot to make your messages worse.",
     "main": "index.js",
     "repository": "https://github.com/artificialbutter/UwuifierBot",
